+++ frontend/js/src/explore/cover-art-collage/2022/CoverArtComposite.tsx
