+++ frontend/js/src/explore/cover-art-collage/2023/CoverArtComposite.tsx
