+++ frontend/js/tests/huesound/ColorPlay.test.tsx
@@ -7,7 +7,7 @@ import { GlobalAppContextT } from "../../src/utils/GlobalAppContext";
 import APIService from "../../src/utils/APIService";
 import BrainzPlayer from "../../src/brainzplayer/BrainzPlayer";
 import * as colorPlayProps from "../__mocks__/colorPlayProps.json";
-import ColorPlay from "../../src/explore/huesound/page";
+import ColorPlay from "../../src/explore/huesound/ColorPlay";
 import ColorWheel from "../../src/explore/huesound/components/ColorWheel";
 import RecordingFeedbackManager from "../../src/utils/RecordingFeedbackManager";
