+++ webpack.config.js
@@ -58,7 +58,7 @@ module.exports = function (env, argv) {
         path.resolve(jsDir, "src/recent/RecentListens.tsx"),
         path.resolve(cssDir, "main.less"),
       ],
-      AIBrainz: [path.resolve(jsDir, "src/explore/ai-brainz/page.tsx")],
+      AIBrainz: [path.resolve(jsDir, "src/explore/ai-brainz/AIBrainz.tsx")],
       listens: [path.resolve(jsDir, "src/user/Listens.tsx")],
       import: path.resolve(jsDir, "src/lastfm/LastFMImporter.tsx"),
       userEntityChart: path.resolve(jsDir, "src/stats/UserEntityChart.tsx"),
@@ -67,28 +67,28 @@ module.exports = function (env, argv) {
       userFeed: path.resolve(jsDir, "src/user-feed/UserFeed.tsx"),
       playlist: path.resolve(jsDir, "src/playlists/Playlist.tsx"),
       playlists: path.resolve(jsDir, "src/playlists/Playlists.tsx"),
-      explore: path.resolve(jsDir, "src/explore/page.tsx"),
-      huesound: path.resolve(jsDir, "src/explore/huesound/page.tsx"),
-      lb_radio: path.resolve(jsDir, "src/explore/lb-radio/page.tsx"),
+      explore: path.resolve(jsDir, "src/explore/Explore.tsx"),
+      huesound: path.resolve(jsDir, "src/explore/huesound/ColorPlay.tsx"),
+      lb_radio: path.resolve(jsDir, "src/explore/lb-radio/LBRadio.tsx"),
       yearInMusic2021: path.resolve(
         jsDir,
-        "src/user/year-in-music/2021/page.tsx"
+        "src/user/year-in-music/2021/YearInMusic.tsx"
       ),
       yearInMusic2022: path.resolve(
         jsDir,
-        "src/user/year-in-music/2022/page.tsx"
+        "src/user/year-in-music/2022/YearInMusic.tsx"
       ),
       yearInMusic2023: path.resolve(
         jsDir,
-        "src/user/year-in-music/2023/page.tsx"
+        "src/user/year-in-music/2023/YearInMusic.tsx"
       ),
       coverArtComposite2022: path.resolve(
         jsDir,
-        "src/explore/cover-art-collage/2022/page.tsx"
+        "src/explore/cover-art-collage/2022/CoverArtComposite.tsx"
       ),
       coverArtComposite2023: path.resolve(
         jsDir,
-        "src/explore/cover-art-collage/2023/page.tsx"
+        "src/explore/cover-art-collage/2023/CoverArtComposite.tsx"
       ),
       homepage: path.resolve(jsDir, "src/home/Homepage.tsx"),
       recommendationsPlayground: path.resolve(
@@ -108,7 +108,10 @@ module.exports = function (env, argv) {
         jsDir,
         "src/metadata-viewer/MetadataViewerPageWrapper.tsx"
       ),
-      freshReleases: path.resolve(jsDir, "src/explore/fresh-releases/page.tsx"),
+      freshReleases: path.resolve(
+        jsDir,
+        "src/explore/fresh-releases/FreshReleases.tsx"
+      ),
       selectTimezone: path.resolve(
         jsDir,
         "src/user-settings/SelectTimezone.tsx"
@@ -117,7 +120,7 @@ module.exports = function (env, argv) {
         jsDir,
         "src/user-settings/SelectTroiPreferences.tsx"
       ),
-      artCreator: path.resolve(jsDir, "src/explore/art-creator/page.tsx"),
+      artCreator: path.resolve(jsDir, "src/explore/art-creator/ArtCreator.tsx"),
       artistPage: path.resolve(jsDir, "src/entity-pages/ArtistPage.tsx"),
       albumPage: path.resolve(jsDir, "src/entity-pages/AlbumPage.tsx"),
     },
